fix test running Signed-off-by: dbulahov <bulahovdenis@gmail.com>
